rename keys.ts to echoProtoJson.ts
