+++ browser-test/fixtures/echoProtoJson.ts
