+++ browser-test/test.grpc.ts
@@ -2,7 +2,7 @@ import * as assert from 'assert';
 import {expect} from 'chai';
 import * as protobuf from 'protobufjs';
 import {GrpcClient} from '../src/browser';
-import {echoProtoJson} from './fixtures/keys';
+import {echoProtoJson} from './fixtures/echoProtoJson';
 
 const authStub = {
   getRequestHeaders() {
